rebase with dashboard changes, move journal store, solve conflicts
