+++ src/app/pages/journal/journal.page.ts
@@ -1,13 +1,333 @@
-import { Component, OnInit } from '@angular/core';
+import {
+  Component, ComponentFactoryResolver, OnInit, ViewChild, ViewContainerRef,
+} from '@angular/core';
+import {
+  LoadingController, IonRefresher,
+  NavParams, Platform, ModalController,
+} from '@ionic/angular';
+import { select, Store } from '@ngrx/store';
+import { Observable, Subscription, merge } from 'rxjs';
+import { map } from 'rxjs/operators';
+
+// import { ScreenOrientation } from '@ionic-native/screen-orientation';
+import { Router } from '@angular/router';
+import {
+  // ActivityCode,
+  SearchResultTestSchema,
+} from '@dvsa/mes-search-schema';
+// TODO Reintroduce this after MES-6242
+// import { TestSlot } from '@dvsa/mes-journal-schema';
+import { BasePageComponent } from '../../shared/classes/base-page';
+import { AuthenticationProvider } from '../../providers/authentication/authentication';
+import * as journalActions from '../../../store/journal/journal.actions';
+import { StoreModel } from '../../shared/models/store.model';
+import {
+  getError, getIsLoading, getSelectedDate, getLastRefreshed,
+  getLastRefreshedTime, getSlotsOnSelectedDate, // getCompletedTests,
+} from '../../../store/journal/journal.selector';
+import { getJournalState } from '../../../store/journal/journal.reducer';
+// TODO Reintroduce this after MES-6242
+// import { SlotSelectorProvider } from '../../providers/slot-selector/slot-selector';
+// import { SlotComponent } from '../../components/test-slot/slot/slot';
+import { SlotItem } from '../../providers/slot-selector/slot-item';
+// TODO Reintroduce this after MES-6419
+// import { getAppInfoState } from '../../modules/app-info/app-info.reducer';
+// import { getVersionNumber } from '../../modules/app-info/app-info.selector';
+import { DateTimeProvider } from '../../providers/date-time/date-time';
+import { AppConfigProvider } from '../../providers/app-config/app-config';
+import { ERROR_PAGE } from '../page-names.constants';
+// import { ErrorTypes } from '../../shared/models/error-message';
+// TODO Reintroduce this after MES-6251
+// import { DeviceProvider } from '../../providers/device/device';
+// import { Insomnia } from '@ionic-native/insomnia';
+// import { PersonalCommitmentSlotComponent } from './personal-commitment/personal-commitment';
+// TODO Reintroduce this after MES-6242
+// import { TestSlotComponent } from '../../components/test-slot/test-slot/test-slot';
+// import { IncompleteTestsBanner } from '../../components/common/incomplete-tests-banner/incomplete-tests-banner';
+import { DateTime } from '../../shared/helpers/date-time';
+import { MesError } from '../../shared/models/mes-error.model';
+// TODO Reintroduce this after MES-6251
+// import { SearchProvider } from '../../providers/search/search';
+// import { formatApplicationReference } from '../../shared/helpers/formatters';
+// import { ApplicationReference } from '@dvsa/mes-test-schema/categories/common';
+import { AppComponent } from '../../app.component';
+// import { isEmpty } from 'lodash';
+// import { TestStatus } from '../../modules/tests/test-status/test-status.model';
+
+interface JournalPageState {
+  selectedDate$: Observable<string>;
+  slots$: Observable<SlotItem[]>;
+  error$: Observable<MesError>;
+  isLoading$: Observable<boolean>;
+  lastRefreshedTime$: Observable<string>;
+  // appVersion$: Observable<string>;
+  // completedTests$: Observable<SearchResultTestSchema[]>;
+}
 
 @Component({
   selector: 'app-journal',
   templateUrl: './journal.page.html',
   styleUrls: ['./journal.page.scss'],
 })
-export class JournalPage implements OnInit {
+export class JournalPage extends BasePageComponent implements OnInit {
 
-  ngOnInit() {
+  @ViewChild('slotContainer', { read: ViewContainerRef }) slotContainer;
+
+  // @ViewChild(IncompleteTestsBanner)
+  // incompleteTestsBanner: IncompleteTestsBanner;
+
+  pageState: JournalPageState;
+  selectedDate: string;
+  // loadingSpinner: Loading;
+  pageRefresher: IonRefresher;
+  isUnauthenticated: boolean;
+  subscription: Subscription;
+  employeeId: string;
+  start = '2018-12-10T08:10:00+00:00';
+  merged$: Observable<void | number>;
+  todaysDate: DateTime;
+  completedTests: SearchResultTestSchema[];
+
+  constructor(
+    public modalController: ModalController,
+    public platform: Platform,
+    public authenticationProvider: AuthenticationProvider,
+    public navParams: NavParams,
+    public loadingController: LoadingController,
+    public router: Router,
+    private store$: Store<StoreModel>,
+    // private slotSelector: SlotSelectorProvider,
+    private resolver: ComponentFactoryResolver,
+    public dateTimeProvider: DateTimeProvider,
+    public appConfigProvider: AppConfigProvider,
+    private app: AppComponent,
+    // TODO Reintroduce this after MES-6251
+    // private deviceProvider: DeviceProvider,
+    // public screenOrientation: ScreenOrientation,
+    // public insomnia: Insomnia,
+    // TODO Reintroduce this after MES-6251
+    // public searchProvider: SearchProvider,
+  ) {
+    super(platform, authenticationProvider, router);
+    this.employeeId = this.authenticationProvider.getEmployeeId();
+    this.isUnauthenticated = this.authenticationProvider.isInUnAuthenticatedMode();
+    this.store$.dispatch(new journalActions.SetSelectedDate(this.dateTimeProvider.now().format('YYYY-MM-DD')));
+    this.todaysDate = this.dateTimeProvider.now();
   }
 
+  ngOnInit(): void {
+    this.pageState = {
+      selectedDate$: this.store$.pipe(
+        select(getJournalState),
+        map(getSelectedDate),
+      ),
+      slots$: this.store$.pipe(
+        select(getJournalState),
+        map(getSlotsOnSelectedDate),
+      ),
+      error$: this.store$.pipe(
+        select(getJournalState),
+        map(getError),
+      ),
+      isLoading$: this.store$.pipe(
+        select(getJournalState),
+        map(getIsLoading),
+      ),
+      lastRefreshedTime$: this.store$.pipe(
+        select(getJournalState),
+        map(getLastRefreshed),
+        map(getLastRefreshedTime),
+      ),
+      // TODO Reintroduce this after MES-6251
+      // appVersion$: this.store$.pipe(
+      //   select(getAppInfoState),
+      //   map(getVersionNumber),
+      // ),
+      // completedTests$: this.store$.pipe(
+      //   select(getJournalState),
+      //   select(getCompletedTests),
+      // ),
+    };
+
+    const {
+      selectedDate$,
+      slots$,
+      error$,
+      // isLoading$,
+      // completedTests$,
+    } = this.pageState;
+
+    this.merged$ = merge(
+      selectedDate$.pipe(map(this.setSelectedDate)),
+      // completedTests$.pipe(map(this.setCompletedTests)),
+      // slots$.pipe(map(this.createSlots)),
+      slots$.pipe(map((value) => console.log(JSON.stringify(value)))),
+      error$.pipe(map(this.showError)),
+      // isLoading$.pipe(map(this.handleLoadingUI)),
+    );
+
+  }
+
+  ionViewDidLeave(): void {
+    // Using .merge helps with unsubscribing
+    if (this.subscription) {
+      this.subscription.unsubscribe();
+    }
+  }
+
+  ionViewWillEnter() {
+    this.loadJournalManually();
+    // this.setupPolling();
+
+    // this.store$.dispatch(new journalActions.LoadCompletedTests());
+
+    if (this.merged$) {
+      this.subscription = this.merged$.subscribe();
+    }
+    this.todaysDate = this.dateTimeProvider.now();
+
+    return true;
+  }
+
+  ionViewWillLeave() {
+    // this.store$.dispatch(new journalActions.StopPolling());
+  }
+
+  ionViewDidEnter(): void {
+    this.store$.dispatch(new journalActions.JournalViewDidEnter());
+
+    if (super.isIos()) {
+      // this.screenOrientation.unlock();
+      // this.insomnia.allowSleepAgain();
+      // this.deviceProvider.disableSingleAppMode();
+    }
+  }
+
+  loadJournalManually() {
+    this.store$.dispatch(new journalActions.LoadJournal());
+  }
+
+  setupPolling() {
+    this.store$.dispatch(new journalActions.SetupPolling());
+  }
+
+  setSelectedDate = (selectedDate: string): void => {
+    this.selectedDate = selectedDate;
+  };
+
+  setCompletedTests = (completedTests: SearchResultTestSchema[]): void => {
+    this.completedTests = completedTests;
+  };
+
+  // handleLoadingUI = (isLoading: boolean): void => {
+  //   if (isLoading) {
+  //     this.loadingSpinner = this.loadingController.create({
+  //       dismissOnPageChange: true,
+  //       spinner: 'circles',
+  //     });
+  //     this.loadingSpinner.present();
+  //     return;
+  //   }
+  //   this.pageRefresher ? this.pageRefresher.complete() : null;
+  //   if (this.loadingSpinner) {
+  //     this.loadingSpinner.dismiss();
+  //     this.loadingSpinner = null;
+  //   }
+  // }
+
+  showError = (error: MesError): void => {
+    if (error === undefined || error.message === '') return;
+
+    // Modals are at the same level as the ion-nav so are not getting the zoom level class,
+    // this needs to be passed in the create options.
+
+    // TODO figure out what to do about zoom methods not existing in appComponent
+    // const zoomClass = `modal-fullscreen ${this.app.getTextZoomClass()}`;
+
+    this.modalController.create({
+      component: ERROR_PAGE,
+      // TODO figure out where this error type is injected now that modal controller structure has changed
+      // type: ErrorTypes.JOURNAL_REFRESH,
+      // cssClass: zoomClass,
+    }).then((modal) => {
+      modal.present();
+    });
+  };
+
+  /**
+   * Returns the activity code if the test has been completed already
+   * Returns null if test hasn't been completed yet
+   */
+  // TODO reintroduce in MES-6242
+  // hasSlotBeenTested(slotData: TestSlot): ActivityCode | null {
+  //   if (isEmpty(this.completedTests)) {
+  //     return null;
+  //   }
+  //
+  //   const applicationReference: ApplicationReference = {
+  //     applicationId: slotData.booking.application.applicationId,
+  //     bookingSequence: slotData.booking.application.bookingSequence,
+  //     checkDigit: slotData.booking.application.checkDigit,
+  //   };
+  //
+  //   const completedTest = this.completedTests.find((completedTest) => {
+  //     return completedTest.applicationReference === parseInt(formatApplicationReference(applicationReference), 10);
+  //   });
+  //
+  //   return completedTest ? completedTest.activityCode : null;
+  // }
+
+  // TODO Reintroduce in MES-6242
+  // private createSlots = (emission: SlotItem[]) => {
+  //   // Clear any dynamically created slots before adding the latest
+  //   this.slotContainer.clear();
+  //
+  //   if (!Array.isArray(emission)) return;
+  //
+  //   if (emission.length === 0) return;
+  //
+  //   const slots = this.slotSelector.getSlotTypes(emission);
+  //
+  //   let lastLocation;
+  //   for (const slot of slots) {
+  //     const factory = this.resolver.resolveComponentFactory(slot.component);
+  //     const componentRef = this.slotContainer.createComponent(factory);
+  //
+  //     (<SlotComponent>componentRef.instance).slot = slot.slotData;
+  //     (<SlotComponent>componentRef.instance).hasSlotChanged = slot.hasSlotChanged;
+  //     (<SlotComponent>componentRef.instance).showLocation = (slot.slotData.testCentre.centreName !== lastLocation);
+  //     lastLocation = slot.slotData.testCentre.centreName;
+  //
+  //     if (componentRef.instance instanceof PersonalCommitmentSlotComponent) {
+  //       // if this is a personal commitment assign it to the component
+  //       (<PersonalCommitmentSlotComponent>componentRef.instance).personalCommitments = slot.personalCommitment;
+  //     }
+  //
+  //     if (componentRef.instance instanceof TestSlotComponent) {
+  //       const activityCode = this.hasSlotBeenTested(slot.slotData as TestSlot);
+  //
+  //       if (activityCode) {
+  //         (<TestSlotComponent>componentRef.instance).derivedActivityCode = activityCode;
+  //         (<TestSlotComponent>componentRef.instance).derivedTestStatus = TestStatus.Submitted;
+  //       }
+  //
+  //       // if this is a test slot assign hasSeenCandidateDetails separately
+  //       (<TestSlotComponent>componentRef.instance).hasSeenCandidateDetails = slot.hasSeenCandidateDetails;
+  //     }
+  //   }
+  // }
+
+  public pullRefreshJournal = (refresher: IonRefresher) => {
+    this.loadJournalManually();
+    this.pageRefresher = refresher;
+  };
+
+  public refreshJournal = () => {
+    this.loadJournalManually();
+  };
+
+  async logout() {
+    this.store$.dispatch(new journalActions.UnloadJournal());
+    await super.logout();
+  }
 }
