+++ src/store/journal/journal.module.ts
@@ -3,13 +3,13 @@ import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
 
 import { RouterModule } from '@angular/router';
-import { DateTimeProvider } from '../../providers/date-time/date-time';
+import { DateTimeProvider } from '../../app/providers/date-time/date-time';
 import { journalReducer } from './journal.reducer';
 import { JournalEffects } from './journal.effects';
 // TODO Re-introduce in MES-6242
 // import { SlotSelectorProvider } from '../../providers/slot-selector/slot-selector';
 // import { SlotProvider } from '../../providers/slot/slot';
-import { JournalProvider } from '../../providers/journal/journal';
+import { JournalProvider } from '../../app/providers/journal/journal';
 import { JournalLogsEffects } from './journal.logs.effects';
 
 @NgModule({
